files/requirements.txt
python-telegram-bot==21.6
groq==0.9.0
